+++ tests/conftest.py
@@ -1,5 +1,6 @@
 """PyTest configuration module. Defines useful fixtures, command line args."""
 import logging
+import shutil
 from pathlib import Path
 
 import numpy as np
@@ -7,7 +8,7 @@ import pandas as pd
 import pytest
 from etoolbox.datazip import DataZip
 
-from dispatch import DispatchModel
+from dispatch import DispatchModel, apply_op_ret_date
 
 logger = logging.getLogger(__name__)
 
@@ -37,6 +38,15 @@ def test_dir() -> Path:
     Mostly this is meant as an example of a fixture.
     """
     return Path(__file__).parent
+
+
+@pytest.fixture(scope="session")
+def temp_dir(test_dir) -> Path:
+    """Return the path to a temp directory that gets deleted on teardown."""
+    out = test_dir / "temp"
+    out.mkdir(exist_ok=True)
+    yield out
+    shutil.rmtree(out)
 
 
 @pytest.fixture(scope="session")
@@ -77,11 +87,13 @@ def ent_redispatch(test_dir) -> dict:
 
 
 @pytest.fixture(scope="session", params=["8fresh", "8redispatch"])
-def ent_dm(test_dir, request) -> DispatchModel:
+def ent_dm(test_dir, request) -> tuple[str, DispatchModel]:
     """Fossil Profiles."""
-    return DispatchModel(
-        **DataZip.dfs_from_zip(test_dir / f"data/{request.param}.zip")
-    )()
+    indicator = {"8fresh": "f", "8redispatch": "r"}
+    return (
+        indicator[request.param],
+        DispatchModel(**DataZip.dfs_from_zip(test_dir / f"data/{request.param}.zip"))(),
+    )
 
 
 @pytest.fixture(scope="session")
@@ -107,3 +119,25 @@ def ent_out_for_test(test_dir):
     self()
     df = self.dispatchable_summary(by=None)
     return df.groupby(level=[0, 1]).sum()
+
+
+@pytest.fixture(scope="session")
+def mini_dm(fossil_profiles, fossil_specs, fossil_cost, re_profiles):
+    """Setup `DispatchModel`."""
+    re = np.array([5000.0, 5000.0, 0.0, 0.0])
+
+    fossil_profiles.columns = fossil_specs.index
+    fossil_profiles = apply_op_ret_date(
+        fossil_profiles, fossil_specs.operating_date, fossil_specs.retirement_date
+    )
+    dm = DispatchModel.from_patio(
+        fossil_profiles.sum(axis=1) - re_profiles @ re,
+        dispatchable_profiles=fossil_profiles,
+        cost_data=fossil_cost,
+        plant_data=fossil_specs,
+        storage_specs=pd.DataFrame(
+            [(5000, 4, 0.9), (2000, 8760, 0.5)],
+            columns=["capacity_mw", "duration_hrs", "roundtrip_eff"],
+        ),
+    )()
+    return dm